Refactor Updatable into its own file
