+++ test/lib/sparse-map-spec.ts
@@ -1,6 +1,6 @@
 import { Observable } from 'rxjs/Observable';
 import { Subject } from 'rxjs/Subject';
-import { Updatable } from '../../src/lib/sparse-map';
+import { Updatable } from '../../src/lib/updatable';
 
 import {expect} from '../support';
